refactor: consolidate 7 riteway.ts copies into one canonical location
Move the shared test assert helper from 7 per-directory __tests__/
copies to a single apps/web/src/test/riteway.ts. Update all 16 test
file imports to use @/test/riteway.

Also update vitest.workspace.ts to reference apps/web/vitest.config.ts
directly so the @/ path alias resolves correctly from the root runner.
