+++ apps/web/src/lib/editor/code-block/__tests__/code-block-shiki.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { Editor } from '@tiptap/core';
 import StarterKit from '@tiptap/starter-kit';
-import { assert } from './riteway';
+import { assert } from '@/test/riteway';
 import { CodeBlockShiki } from '../CodeBlockShikiExtension';
 
 function createEditor(content?: string) {
