+++ apps/web/src/lib/integrations/google-calendar/__tests__/map-attendees.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, vi, beforeEach } from 'vitest';
-import { assert } from './riteway';
+import { assert } from '@/test/riteway';
 
 // Track calls for assertions
 const mockSelectFrom = vi.fn().mockReturnValue({ where: vi.fn().mockResolvedValue([]) });
