+++ apps/web/src/lib/integrations/google-calendar/__tests__/sync-service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, vi, beforeEach } from 'vitest';
-import { assert } from './riteway';
+import { assert } from '@/test/riteway';
 
 // Mock database
 const mockFindFirst = vi.fn();
