+++ apps/web/src/lib/memory/__tests__/discovery-service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, vi, beforeEach } from 'vitest';
-import { assert } from './riteway';
+import { assert } from '@/test/riteway';
 
 /**
  * Discovery Service Tests
