+++ apps/web/src/lib/utils/__tests__/timestamp.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from 'vitest';
-import { assert } from './riteway';
+import { assert } from '@/test/riteway';
 import { toISOTimestamp } from '../timestamp';
 
 describe('toISOTimestamp', () => {
